build: DTO 타입 변경

+++ app/src/main/java/com/teamwss/websoso/data/remote/response/NovelMemoInfoResponse.kt
@@ -4,7 +4,7 @@ data class NovelMemoInfoResponse(
     val memos: List<NovelMemoResponse>,
     val platforms: List<PlatformGetResponse>,
     val userNovelGenre: String,
-    val userNovelRating: Double,
+    val userNovelRating: Float,
     val userNovelReadEndDate: String,
     val userNovelReadStartDate: String,
     val userNovelReadStatus: String,
